+++ app/models/concerns/events/__init__.py
@@ -1,0 +1,5 @@
+from .attendable import Attendable
+from .collaborative import Collaborative
+
+
+__all__ = ["Attendable", "Collaborative"]
